Justeret integer-variable så de passer til array.size() som nu returnerer size_t som er unsigned int i x86 og unsigned __int64 i x64 tilføjet In_z_ _Printf_format_string_ annotation de fleste steder hvor der bruges format-strenge. Virker ikke!! bruger std:streamsize for parametrene til StrStream/StrStreamParameters, spaceString Tilføjet2 inline functioner:htonll/ntohll som konverterer __int64 ligsom htonl og ntohl gør for unsigned long Tilføjet ny type i Packer.ElementType:E_LONG_LONG

+++ Lib/Include/BigArray.h
@@ -574,7 +574,7 @@ public:
       } else {
         BitSet expectedFreeFilePages(m_freeFilePages.getCapacity());
         expectedFreeFilePages.add(0, m_pageFileSize-1);
-        for(int i = 0; i < m_pageMap.size(); i++) {
+        for(size_t i = 0; i < m_pageMap.size(); i++) {
           const PageMapElement &pm = m_pageMap[i];
           if(pm.m_loaded == 0) {
             expectedFreeFilePages.remove(pm.m_pageIndex);
